+++ CopilotKit/packages/runtime/src/lib/runtime/copilot-runtime.ts
@@ -39,8 +39,8 @@ import { Message } from "../../graphql/types/converted";
 import { ForwardedParametersInput } from "../../graphql/inputs/forwarded-parameters.input";
 
 import {
-  isLangGraphAgentAction,
-  LangGraphAgentAction,
+  isRemoteAgentAction,
+  RemoteAgentAction,
   EndpointType,
   setupRemoteActions,
   EndpointDefinition,
@@ -153,7 +153,7 @@ export interface CopilotRuntimeConstructorParams<T extends Parameter[] | [] = []
   middleware?: Middleware;
 
   /*
-   * A list of server side actions that can be executed.
+   * A list of server side actions that can be executed. Will be ignored when remoteActions are set
    */
   actions?: ActionsConfiguration<T>;
 
@@ -190,7 +190,13 @@ export class CopilotRuntime<const T extends Parameter[] | [] = []> {
   private delegateAgentProcessingToServiceAdapter: boolean;
 
   constructor(params?: CopilotRuntimeConstructorParams<T>) {
-    this.actions = params?.actions || [];
+    // Do not register actions if endpoints are set
+    if (params?.actions && params?.remoteEndpoints) {
+      console.warn("Actions set in runtime instance will be ignored when remote endpoints are set");
+      this.actions = [];
+    } else {
+      this.actions = params?.actions || [];
+    }
 
     for (const chain of params?.langserve || []) {
       const remoteChain = new RemoteChain(chain);
@@ -299,7 +305,7 @@ please use an LLM adapter instead.`,
           (action) =>
             // TODO-AGENTS: do not exclude ALL server side actions
             !serverSideActions.find((serverSideAction) => serverSideAction.name == action.name),
-          // !isLangGraphAgentAction(
+          // !isRemoteAgentAction(
           //   serverSideActions.find((serverSideAction) => serverSideAction.name == action.name),
           // ),
         ),
@@ -316,7 +322,6 @@ please use an LLM adapter instead.`,
   }
 
   async discoverAgentsFromEndpoints(graphqlContext: GraphQLContext): Promise<AgentWithEndpoint[]> {
-    const headers = createHeaders(null, graphqlContext);
     const agents = this.remoteEndpointDefinitions.reduce(
       async (acc: Promise<Agent[]>, endpoint) => {
         const agents = await acc;
@@ -349,12 +354,12 @@ please use an LLM adapter instead.`,
             description: string;
           }>;
         }
-
-        const fetchUrl = `${(endpoint as CopilotKitEndpoint).url}/info`;
+        const cpkEndpoint = endpoint as CopilotKitEndpoint;
+        const fetchUrl = `${endpoint.url}/info`;
         try {
           const response = await fetch(fetchUrl, {
             method: "POST",
-            headers,
+            headers: createHeaders(cpkEndpoint.onBeforeRequest, graphqlContext),
             body: JSON.stringify({ properties: graphqlContext.properties }),
           });
           if (!response.ok) {
@@ -438,11 +443,12 @@ please use an LLM adapter instead.`,
       agentWithEndpoint.endpoint.type === EndpointType.CopilotKit ||
       !("type" in agentWithEndpoint.endpoint)
     ) {
-      const fetchUrl = `${(agentWithEndpoint.endpoint as CopilotKitEndpoint).url}/agents/state`;
+      const cpkEndpoint = agentWithEndpoint.endpoint as CopilotKitEndpoint;
+      const fetchUrl = `${cpkEndpoint.url}/agents/state`;
       try {
         const response = await fetch(fetchUrl, {
           method: "POST",
-          headers,
+          headers: createHeaders(cpkEndpoint.onBeforeRequest, graphqlContext),
           body: JSON.stringify({
             properties: graphqlContext.properties,
             threadId,
@@ -499,8 +505,8 @@ please use an LLM adapter instead.`,
     const messages = convertGqlInputToMessages(rawMessages);
 
     const currentAgent = serverSideActions.find(
-      (action) => action.name === agentName && isLangGraphAgentAction(action),
-    ) as LangGraphAgentAction;
+      (action) => action.name === agentName && isRemoteAgentAction(action),
+    ) as RemoteAgentAction;
 
     if (!currentAgent) {
       throw new CopilotKitAgentDiscoveryError({ agentName });
@@ -513,9 +519,9 @@ please use an LLM adapter instead.`,
       .filter(
         (action) =>
           // Case 1: Keep all regular (non-agent) actions
-          !isLangGraphAgentAction(action) ||
+          !isRemoteAgentAction(action) ||
           // Case 2: For agent actions, keep all except self (prevent infinite loops)
-          (isLangGraphAgentAction(action) && action.name !== agentName) /* prevent self-calls */,
+          (isRemoteAgentAction(action) && action.name !== agentName) /* prevent self-calls */,
       )
       .map((action) => ({
         name: action.name,
@@ -536,7 +542,7 @@ please use an LLM adapter instead.`,
     });
     try {
       const eventSource = new RuntimeEventSource();
-      const stream = await currentAgent.langGraphAgentHandler({
+      const stream = await currentAgent.remoteAgentHandler({
         name: agentName,
         threadId,
         nodeName,
@@ -572,7 +578,7 @@ please use an LLM adapter instead.`,
         threadId,
         runId: undefined,
         eventSource,
-        serverSideActions: [],
+        serverSideActions,
         actionInputsWithoutAgents: allAvailableActions,
       };
     } catch (error) {
@@ -651,7 +657,7 @@ export function langGraphPlatformEndpoint(
 
 export function resolveEndpointType(endpoint: EndpointDefinition) {
   if (!endpoint.type) {
-    if ("langsmithApiKey" in endpoint && "deploymentUrl" in endpoint && "agents" in endpoint) {
+    if ("deploymentUrl" in endpoint && "agents" in endpoint) {
       return EndpointType.LangGraphPlatform;
     } else {
       return EndpointType.CopilotKit;